fix bugs and add the log register part, add also the ui to the log reports

+++ sublicreacr.Negocio2/Empresa.cs
@@ -8,14 +8,14 @@ namespace sublicreacr.Negocio
 {
     public class Empresa
     {
-        private int cedulaJuridica;
+        private long cedulaJuridica;
         private string nombreEmpresa;
-        private int telefono;
+        private long telefono;
         private byte[] logo;
         private bool estado;
         private string estadoLeyenda;
 
-        public Empresa(int _cedulaJuridica,string _nombreEmpresa,int _telefono,byte[] _logo
+        public Empresa(long _cedulaJuridica,string _nombreEmpresa,long _telefono,byte[] _logo
             , bool _estado, string _estadoLeyenda)
         {
             this.cedulaJuridica = _cedulaJuridica;
@@ -29,9 +29,9 @@ namespace sublicreacr.Negocio
         public Empresa() 
         { 
         }
-        public int CedulaJuridica { get=>cedulaJuridica; set=>cedulaJuridica=value; }
+        public long CedulaJuridica { get=>cedulaJuridica; set=>cedulaJuridica=value; }
         public string NombreEmpresa { get => nombreEmpresa; set => nombreEmpresa = value; }
-        public int Telefono { get => telefono; set => telefono = value; }
+        public long Telefono { get => telefono; set => telefono = value; }
         public byte[] Logo { get => logo; set => logo = value; }
         public bool Estado { get => estado; set => estado = value; }
         public string EstadoLeyenda { get => estadoLeyenda; set => estadoLeyenda = value; }
